requirements/base.txt
Django==4.1.7
djangorestframework==3.14.0
django-ckeditor==6.5.1
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.9.0
django-filter==22.1
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
Pillow==9.4.0
pytest-django==4.5.2
psycopg2-binary==2.9.5
